[FEATURE][TAS-134] Removed event end date, allow add category on event creation
