+++ events/admin.py
@@ -81,8 +81,7 @@ class EventAdmin(admin.ModelAdmin):
                     'event_source_url',
                     'image',
                     'image_source_url',
-                    ('category',),
-                    ('event_date', 'event_end_date'),
+                    ('event_date', 'category'),
                     'price',
                     'organizers',
                     'place',
